Adicionado o package tile. A classe Tile foi movida para o package tile. Nova classe criada Solid para a verificação de colisão com Tiles. As variáveis x e y de Player foram alteradas de double para int. Adicionado um atributo constante TILE_SIZE = 16 para facilitar o entendimento e a realização de alterações no código. O atributo

+++ src/com/zelda/entity/Avatar.java
@@ -7,9 +7,9 @@ import java.awt.image.BufferedImage;
 
 public class Avatar {
 	
-	protected double x;
+	protected int x;
 	
-	protected double y;
+	protected int y;
 	
 	protected int width;
 	
@@ -33,11 +33,11 @@ public class Avatar {
 		graphics.drawImage(getSprite(), (int)getX() - Camera.x, (int)getY() - Camera.y, null);
 	}
 	
-	private void setX(double x) {
+	private void setX(int x) {
 		this.x = x;
 	}
 	
-	private void setY(double y) {
+	private void setY(int y) {
 		this.y = y;
 	}
 	
@@ -53,11 +53,11 @@ public class Avatar {
 		this.sprite = sprite;
 	}
 	
-	public double getX() {
+	public int getX() {
 		return this.x;
 	}
 	
-	public double getY() {
+	public int getY() {
 		return this.y;
 	}
